refactor: add label renderers
