+++ client/src/modules/Schedule/components/StatusTabs/renderers.tsx
@@ -1,20 +1,7 @@
-import { Space } from 'antd';
-import { ReactNode } from 'react';
 import { CourseScheduleItemDtoStatusEnum } from 'api';
 import { ALL_TAB_KEY, ALL_TAB_LABEL, SCHEDULE_STATUSES } from 'modules/Schedule/constants';
 import { Status } from './StatusTabs';
-import { CountBadge } from 'components/CountBadge';
-
-type Item = {
-  label: string;
-  key: string;
-  count: number;
-};
-
-type TabItem = {
-  label: ReactNode;
-  key: string;
-};
+import { LabelItem, labelRender } from 'components/TabsWithCounter/renderers';
 
 const tabsOrder = [
   ALL_TAB_KEY,
@@ -26,7 +13,7 @@ const tabsOrder = [
   CourseScheduleItemDtoStatusEnum.Archived,
 ];
 
-export const tabsRenderer = (statuses: Status[], activeTab?: string): TabItem[] => {
+export const tabsRenderer = (statuses: Status[], activeTab?: string) => {
   const initialItem = {
     label: ALL_TAB_LABEL,
     key: ALL_TAB_KEY,
@@ -35,15 +22,15 @@ export const tabsRenderer = (statuses: Status[], activeTab?: string): TabItem[] 
 
   return SCHEDULE_STATUSES.reduce(
     (
-      acc: Item[],
+      acc: LabelItem[],
       current: {
         value: CourseScheduleItemDtoStatusEnum;
         text: string;
       },
-    ): Item[] => {
+    ) => {
       const { text, value } = current;
 
-      const newItem: Item = {
+      const newItem = {
         label: text,
         key: value,
         count: statuses.filter(el => el === value).length,
@@ -53,18 +40,5 @@ export const tabsRenderer = (statuses: Status[], activeTab?: string): TabItem[] 
     [initialItem],
   )
     .sort((prev, next) => tabsOrder.indexOf(prev.key) - tabsOrder.indexOf(next.key))
-    .map(({ count, key, label }) => {
-      const isTabActive = activeTab === key;
-      return {
-        key,
-        label: count ? (
-          <Space>
-            {label}
-            <CountBadge count={count} status={isTabActive ? 'processing' : 'default'} />
-          </Space>
-        ) : (
-          label
-        ),
-      };
-    });
+    .map(item => labelRender(item, activeTab));
 };
